Setup: Create modular project structure (#3)
Sets up the project structure by creating files to separate concerns
and adhere to the single responsibility principle. Prefaces each file 
with the module name `httplite` and standardizes the format of the 
initial exploratory code.

At this point in the project, we have set up a few of the core features
we need to build the layer 6 load balancer. Currently, we have set
up the configuration module, an HTTP server, and HTTP requests.

Additionally, we set up structures and lays out the groundwork for
future development as well as scripts for compiling and running our 
module.