Implement: Get request length
Added a function that looks at a slab's buffer containing the request
headers in order to get the total length of the request (headers + 
body). Finds the Content Length header and header/body 
separator to do so. Uses default string search functions; later on
we hope to optimize this to use faster parsing methods.