Implement: Keep-alive for upstreams
This pull request adds a keep-alive timer to the upstreams and allows the server
to destroy any idle connections and conserve resources. The keep-alive is global
for all upstreams and is set in the config file (or with a default value of 30 
seconds). Additionally, some error-checking and error responses have been added
to the code. There are some minor issues that we will attempt to fix later.